% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{audioRecordButtons}
\alias{audioRecordButtons}
\title{Buttons to manage audio recording}
\usage{
audioRecordButtons(id, label_start = "Record", label_stop = "Stop")
}
\arguments{
\item{id}{Module's id.}

\item{label_start}{Label for the start button.}

\item{label_stop}{Label for the stop button.}
}
\description{
Buttons to manage audio recording
}
