% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{renderAudio}
\alias{renderAudio}
\title{Render an audio player output}
\usage{
renderAudio(audio_reactive)
}
\arguments{
\item{audio_reactive}{\code{list}. audioRecordServer returns as reactive.}
}
\description{
Returns a renderUI with audio tags
}
