% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run App}
\usage{
run_app(host = "0.0.0.0", port = 3838)
}
\arguments{
\item{host}{\code{character}. See \code{\link{shinyAppDir}} (options).}

\item{port}{\code{integer}. See \code{\link{shinyAppDir}} (options)..}
}
\description{
Run App
}
