% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runDemo}
\alias{runDemo}
\title{Run App}
\usage{
runDemo(host = "0.0.0.0", port = 3838)
}
\arguments{
\item{host}{\code{character}. See \code{\link{shinyAppDir}} (options).}

\item{port}{\code{integer}. See \code{\link{shinyAppDir}} (options)..}
}
\description{
Run App
}
