% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{writeMP3}
\alias{writeMP3}
\title{Write a MP3 file}
\usage{
writeMP3(audio, path)
}
\arguments{
\item{audio}{\code{list}. evaluated audioRecordServer returns.}

\item{path}{\code{character}. where to save file}
}
\description{
Write a MP3 file from the returns of audioRecordServer
}
